backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0
google-generativeai

# AI API clients (uncomment and install what you need)
# openai==1.3.0
# google-generativeai==0.3.0
# anthropic==0.7.0
